signed/unsigned mismatch fix

+++ Verklegt2014/mainwindow.cpp
@@ -102,7 +102,7 @@ void MainWindow::on_addScientist_clicked()
 void MainWindow::on_removeScientist_clicked()
 {
     auto list = ui->scientistTable->selectedItems();
-    for(size_t i = 0; i < list.size(); i++)
+    for(int i = 0; i < list.size(); i++)
     {
         QTableWidgetItem* item = list.at(i);
     }
